\docType{package}
\name{R2admb-package}
\alias{R2admb}
\alias{R2admb-package}
\title{ADMB to R interface functions}
\description{
  A series of functions to call AD Model Builder (i.e.,
  compile and run models) from within R, read the results
  back into R as "admb" objects, and provide standard
  accessors (i.e. coef(), vcov(), etc.)
}
\details{
  \tabular{ll}{ Package: \tab R2admb\cr Type: \tab
  Package\cr Version: \tab 0.5\cr Date: \tab 2009-11-11\cr
  License: \tab GPL\cr LazyLoad: \tab yes\cr } More here!
}
\author{
  Ben Bolker

  Maintainer: Ben Bolker <bolker@ufl.edu>
}
\references{
  \url{admb-project.org}
}
\seealso{
  PBSadmb package, glmmADMB package, ADMB2R
}
\keyword{package}

