% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.r
\name{admb_version}
\alias{admb_version}
\title{Query ADMB version}
\usage{
admb_version()
}
\value{
Prints the version string from a compiled ADMB file, and returns the value (invisibly) as a character vector; returns NA if ADMB is not installed
}
\description{
Report on the version of ADMB being used.
}
\examples{
admb_version()
}
\author{
Ben Bolker
}
\keyword{misc}
